userland: Add a retry mechanism to fix_TEE()
Depending on when the driver starts after boot, there could be a race
with userspace making the driver unable to access the resetprop binary.
If the driver starts around the 5 seconds after boot mark, the race only
happens on first kernel boot. If for some reason the driver starts very
early at around the 2 seconds after boot mark, it's too early to access
the resetprop binary. In that case, just keep trying until it works.
Even if the command succeeded, if it happened after boot is complete,
the system would have already read the device model props, for that reason
we can't wait for the boot to complete for a sure access to the file.

Signed-off-by: celtare21 <celtare21@gmail.com>